mavlink: move OPTICAL_FLOW to streams hpp
